Refactor Message hierarchy. Compiling, but tests failing.
