--- src/main/java/com/nexmo/client/messaging/NexmoSmsClientSignedRequests.java
@@ -1,72 +1,0 @@
-/*
- * Copyright (c) 2011-2017 Nexmo Inc
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package com.nexmo.client.messaging;
-
-
-/**
- * Client for talking to the Nexmo SMS API that signs requests using a MD5 signature derived from a secret key.
- * <p>
- * For usage information see {@link com.nexmo.client.messaging.NexmoSmsClient}.
- *
- * @author  Paul Cook
- */
-public class NexmoSmsClientSignedRequests extends NexmoSmsClient {
-
-    /**
-     * Instantiate a new NexmoSmsClientSignedRequests instance that will communicate using the supplied credentials.
-     *
-     * @param apiKey Your Nexmo account api key
-     * @param signatureSecretKey Your Nexmo account secret key for signing api requests
-     */
-    public NexmoSmsClientSignedRequests(final String apiKey,
-                                        final String signatureSecretKey) {
-        super(DEFAULT_BASE_URL,
-             apiKey,
-             null,   // password,
-             DEFAULT_CONNECTION_TIMEOUT,
-             DEFAULT_SO_TIMEOUT,
-             true,   // signRequests
-             signatureSecretKey);
-    }
-
-    /**
-     * Instantiate a new NexmoSmsClientSignedRequests instance that will communicate using the supplied credentials, and will use the supplied connection and read timeout values.
-     *
-     * @param apiKey Your Nexmo account api key
-     * @param signatureSecretKey Your Nexmo account secret key for signing api requests
-     * @param connectionTimeout over-ride the default connection timeout with this value (in milliseconds)
-     * @param soTimeout over-ride the default read-timeout with this value (in milliseconds)
-     */
-    public NexmoSmsClientSignedRequests(final String apiKey,
-                                        final String signatureSecretKey,
-                                        final int connectionTimeout,
-                                        final int soTimeout) {
-        super(DEFAULT_BASE_URL,
-             apiKey,
-             null,   // password,
-             connectionTimeout,
-             soTimeout,
-             true,   // signRequests
-             signatureSecretKey);
-    }
-
-}
