+++ src/main/java/com/nexmo/client/messaging/messages/BinaryMessage.java
@@ -22,96 +22,78 @@
 package com.nexmo.client.messaging.messages;
 
 
-import com.nexmo.client.messaging.messages.parameters.MessageClass;
+import com.nexmo.common.util.HexUtil;
+import org.apache.http.client.methods.RequestBuilder;
 
 /**
- * Represents the details of a binary message that is to be submitted via the Nexmo REST API.
- *
- * @author  Paul Cook
+ * A binary message to be submitted via the Nexmo SMS API.
  */
 public class BinaryMessage extends Message {
-
-    private static final long serialVersionUID = -5301775803396025570L;
+    private final byte[] messageBody;
+    private final byte[] udh;
+    private int protocolId = 0;
 
     /**
      * Instantiate a new binary sms message request.
      *
-     * @param from the 'from' address that will be seen on the handset when this message arrives,
-     *             typically either a valid short-code / long code that can be replied to, or a short text description of the application sending the message (Max 11 chars)
-     * @param to   the phone number of the handset that you wish to send the message to
-     * @param binaryMessageBody The raw binary message data to be sent to a handset.
+     * @param from              the 'from' address that will be seen on the handset when this message arrives,
+     *                          typically either a valid short-code / long code that can be replied to, or a short text description of the application sending the message (Max 11 chars)
+     * @param to                the phone number of the handset that you wish to send the message to
+     * @param messageBody The raw binary message data to be sent to a handset.
      *                          This api, and the Nexmo messaging service will send this data 'as-is' (in conjunction with the binary UDH) and will not make any corrections.
      *                          so you should ensure that it is a correctly constructed message
-     * @param binaryMessageUdh Most binary content will require a UserDataHeader portion of the message containing commands to enable the handset to interpret the binary data
-     *                         (for example, a binary ringtone, a wap-push, OverTheAir configuration, etc).
-     *                         Additionally, if you are sending a long text message as multiple concatenated messages and are performing this operation manually rather than
-     *                         using the automated long messaging handling in the Nexmo messaging service, then you will need to construct and include here an appropriate
-     *                         UserDataHeader field that describes the segmentation/re-assembly fields required to successfully concatenate multiple short messages.
+     * @param udh  Most binary content will require a UserDataHeader portion of the message containing commands to enable the handset to interpret the binary data
+     *                          (for example, a binary ringtone, a wap-push, OverTheAir configuration, etc).
+     *                          Additionally, if you are sending a long text message as multiple concatenated messages and are performing this operation manually rather than
+     *                          using the automated long messaging handling in the Nexmo messaging service, then you will need to construct and include here an appropriate
+     *                          UserDataHeader field that describes the segmentation/re-assembly fields required to successfully concatenate multiple short messages.
      */
     public BinaryMessage(final String from,
                          final String to,
-                         final byte[] binaryMessageBody,
-                         final byte[] binaryMessageUdh) {
-        super(MESSAGE_TYPE_BINARY,
-              from,
-              to,
-              null,
-              binaryMessageBody,
-              binaryMessageUdh,
-              null,
-              false,
-              false,
-              null,
-              null,
-              0,
-              null,
-              null);
+                         final byte[] messageBody,
+                         final byte[] udh
+    ) {
+        super(MessageType.BINARY, from, to);
+        this.messageBody = messageBody;
+        this.udh = udh;
     }
 
     /**
-     * Instantiate a new binary sms message request, attaching an optional client reference, message-class or TTL, and optionally requesting a delivery notification.
-     *
-     * @param from the 'from' address that will be seen on the handset when this message arrives,
-     *             typically either a valid short-code / long code that can be replied to, or a short text description of the application sending the message (Max 11 chars)
-     * @param to   the phone number of the handset that you wish to send the message to
-     * @param binaryMessageBody The raw binary message data to be sent to a handset.
-     *                          This api, and the Nexmo messaging service will send this data 'as-is' (in conjunction with the binary UDH) and will not make any corrections.
-     *                          so you should ensure that it is a correctly constructed message
-     * @param binaryMessageUdh Most binary content will require a UserDataHeader portion of the message containing commands to enable the handset to interpret the binary data
-     *                         (for example, a binary ringtone, a wap-push, OverTheAir configuration, etc).
-     *                         Additionally, if you are sending a long text message as multiple concatenated messages and are performing this operation manually rather than
-     *                         using the automated long messaging handling in the Nexmo messaging service, then you will need to construct and include here an appropriate
-     *                         UserDataHeader field that describes the segmentation/re-assembly fields required to successfully concatenate multiple short messages.
-     * @param clientReference This is a user definable reference that will be stored in the Nexmo messaging records. It will be available in detailed reporting / analytics
-     *                        In order to help with reconciliation of messages
-     * @param statusReportRequired If set to true, then a delivery notification will be requested for this message delivery attempt.
-     *                             Upon receiving notification of delivery or failure from the network, the Nexmo platform will submit a notification to the url configured in your
-     *                             Nexmo REST account that represents the outcome of this message.
-     * @param messageClass (Optional) The message class that is to be applied to this message.
-     * @param protocolId The value of the GSM Protocol ID field to be submitted with this message. Ordinarily this should be left as the default value of 0
+     * @return byte[] The raw binary message data to be sent to a handset.
+     * This api, and the Nexmo messaging service will send this data 'as-is' (in conjunction with the binary UDH) and will not make any corrections.
+     * so you should ensure that it is a correctly constructed message
      */
-    public BinaryMessage(final String from,
-                         final String to,
-                         final byte[] binaryMessageBody,
-                         final byte[] binaryMessageUdh,
-                         final String clientReference,
-                         final boolean statusReportRequired,
-                         final MessageClass messageClass,
-                         final Integer protocolId) {
-        super(MESSAGE_TYPE_BINARY,
-              from,
-              to,
-              null,
-              binaryMessageBody,
-              binaryMessageUdh,
-              clientReference,
-              false,
-              statusReportRequired,
-              null,
-              null,
-              0,
-              messageClass,
-              protocolId);
+    public byte[] getMessageBody() {
+        return this.messageBody == null ? null : this.messageBody.clone();
     }
 
+    /**
+     * @return byte[] Most binary content will require a UserDataHeader portion of the message containing commands to enable the handset to interpret the binary data
+     * (for example, a binary ringtone, a wap-push, OverTheAir configuration, etc).
+     * Additionally, if you are sending a long text message as multiple concatenated messages and are performing this operation manually rather than
+     * using the automated long messaging handling in the Nexmo messaging service, then you will need to construct and include here an appropriate
+     * UserDataHeader field that describes the segmentation/re-assembly fields required to successfully concatenate multiple short messages.
+     */
+    public byte[] getUdh() {
+        return this.udh == null ? null : this.udh.clone();
+    }
+
+    /**
+     * @return Integer The value of the GSM Protocol ID field to be submitted with this message. Ordinarily this should be left as the default value of 0
+     */
+    public int getProtocolId() {
+        return this.protocolId;
+    }
+
+    public void setProtocolId(int protocolId) {
+        this.protocolId = protocolId;
+    }
+
+    @Override
+    public void addParams(RequestBuilder request) {
+        super.addParams(request);
+        request.addParameter("udh", HexUtil.bytesToHex(getUdh()));
+        request.addParameter("body", HexUtil.bytesToHex(getMessageBody()));
+        request.addParameter("protocol-id", Integer.toString(protocolId));
+    }
 }
