+++ src/main/java/com/nexmo/client/messaging/messages/TextMessage.java
@@ -22,82 +22,81 @@
 package com.nexmo.client.messaging.messages;
 
 
-import com.nexmo.client.messaging.messages.parameters.MessageClass;
+import org.apache.http.client.methods.RequestBuilder;
 
 /**
  * Represents the details of a plain-text message that is to be submitted via the Nexmo REST api.
  *
- * @author  Paul Cook
+ * @author Paul Cook
  */
 public class TextMessage extends Message {
+    private final String messageBody;
 
-    private static final long serialVersionUID = 6258872793039443129L;
+    private boolean unicode = false;
 
     /**
      * Instantiate a new text-message request.<br>
      * This message will be submitted as a regular 8 bit text message
      *
-     * @param from the 'from' address that will be seen on the handset when this message arrives,
-     *             typically either a valid short-code / long code that can be replied to, or a short text description of the application sending the message (Max 11 chars)
-     * @param to   the phone number of the handset that you wish to send the message to
+     * @param from        the 'from' address that will be seen on the handset when this message arrives,
+     *                    typically either a valid short-code / long code that can be replied to, or a short text
+     *                    description of the application sending the message (Max 11 chars)
+     * @param to          the phone number of the handset that you wish to send the message to
      * @param messageBody The text of the message to be sent to the handset
      */
     public TextMessage(final String from,
                        final String to,
                        final String messageBody) {
-        super(MESSAGE_TYPE_TEXT,
-              from,
-              to,
-              messageBody,
-              null,
-              null,
-              null,
-              false,
-              false,
-              null,
-              null,
-              0,
-              null,
-              null);
+        this(from, to, messageBody, false);
     }
 
     /**
-     * Instantiate a new text-message request, exposing all of the available parameters, and optionally requesting a delivery notification.<br>
+     * Instantiate a new text-message request.<br>
      * This message will be submitted as a regular 8 bit text message
      *
-     * @param from the 'from' address that will be seen on the handset when this message arrives,
-     *             typically either a valid short-code / long code that can be replied to, or a short text description of the application sending the message (Max 11 chars)
-     * @param to   the phone number of the handset that you wish to send the message to
+     * @param from        the 'from' address that will be seen on the handset when this message arrives,
+     *                    typically either a valid short-code / long code that can be replied to, or a short text
+     *                    description of the application sending the message (Max 11 chars)
+     * @param to          the phone number of the handset that you wish to send the message to
      * @param messageBody The text of the message to be sent to the handset
-     * @param clientReference This is a user definable reference that will be stored in the Nexmo messaging records. It will be available in detailed reporting / analytics
-     *                        In order to help with reconciliation of messages
-     * @param statusReportRequired If set to true, then a delivery notification will be requested for this message delivery attempt.
-     *                             Upon receiving notification of delivery or failure from the network, the Nexmo platform will submit a notification to the url configured in your
-     *                             Nexmo REST account that represents the outcome of this message.
-     * @param messageClass (Optional) The message class that is to be applied to this message.
-     * @param protocolId The value of the GSM Protocol ID field to be submitted with this message. Ordinarily this should be left as the default value of 0
      */
     public TextMessage(final String from,
                        final String to,
                        final String messageBody,
-                       final String clientReference,
-                       final boolean statusReportRequired,
-                       final MessageClass messageClass,
-                       final Integer protocolId) {
-        super(MESSAGE_TYPE_TEXT,
-              from,
-              to,
-              messageBody,
-              null,
-              null,
-              clientReference,
-              false,
-              statusReportRequired,
-              null,
-              null,
-              0,
-              messageClass,
-              protocolId);
+                       final boolean unicode) {
+        super(null, from, to);
+        this.messageBody = messageBody;
+        this.unicode = unicode;
     }
 
+    /**
+     * @return String The text of the message to be sent to the handset
+     */
+    public String getMessageBody() {
+        return this.messageBody;
+    }
+
+    /**
+     * @return boolean This flag is set to true if the message needs to be submitted as a unicode message. This would
+     * be for scenario's where the message contains text that does not fit within the Latin GSM alphabet. Examples
+     * would be messages to be sent in non-western scripts, such as Arabic, Kanji, Chinese, etc.
+     */
+    public boolean isUnicode() {
+        return this.unicode;
+    }
+
+    @Override
+    public MessageType getType() {
+        if (unicode) {
+            return MessageType.UNICODE;
+        } else {
+            return MessageType.TEXT;
+        }
+    }
+
+    @Override
+    public void addParams(RequestBuilder request) {
+        super.addParams(request);
+        request.addParameter("text", messageBody);
+    }
 }
