+++ src/main/java/com/nexmo/client/messaging/messages/parameters/MessageClass.java
@@ -32,22 +32,22 @@ public enum MessageClass {
     /**
      * Message Class 0
      */
-    MESSAGE_CLASS_0(0),
+    CLASS_0(0),
 
     /**
      * Message Class 1
      */
-    MESSAGE_CLASS_1(1),
+    CLASS_1(1),
 
     /**
      * Message Class 2
      */
-    MESSAGE_CLASS_2(2),
+    CLASS_2(2),
 
     /**
      * Message Class 3
      */
-    MESSAGE_CLASS_3(3);
+    CLASS_3(3);
 
     private final int messageClass;
